requirements.txt
pyspark==3.4.0
apache-airflow-providers-apache-spark
boto3
nba_api
requests
